+++ offchain/__init__.py
@@ -1,3 +1,7 @@
-from offchain.metadata.models.token import Token  # noqa
-from offchain.metadata.pipelines.metadata_pipeline import \
-    MetadataPipeline  # noqa
+from offchain.metadata import (
+    Metadata,
+    MetadataFetcher,
+    MetadataPipeline,
+    MetadataProcessingError,
+    Token,
+)  # noqa
