+++ offchain/metadata/__init__.py
@@ -1,1 +1,5 @@
-from . import adapters, fetchers, models, parsers
+from .adapters import *  # noqa
+from .fetchers import *  # noqa
+from .models import *  # noqa
+from .parsers import *  # noqa
+from .pipelines import *  # noqa
