+++ offchain/metadata/adapters/__init__.py
@@ -1,5 +1,5 @@
 from .arweave import ARWeaveAdapter
-from .base_adapter import BaseAdapter
+from .base_adapter import AdapterConfig, BaseAdapter
 from .data_uri import DataURIAdapter
 from .http_adapter import HTTPAdapter
 from .ipfs import IPFSAdapter
