+++ offchain/metadata/pipelines/__init__.py
@@ -1,0 +1,2 @@
+from .base_pipeline import BasePipeline
+from .metadata_pipeline import MetadataPipeline
